confetti-and-fixes

+++ src/services/API/httpRequests.js
@@ -30,12 +30,14 @@ export const getTopic = async (topic_id) => {
 export const getUserPlanner = async (user, subject) => {
 
   if (user) {
+    if (!user.isAnonymous) {
       const userToken = await user.getIdToken()
       console.log('getting user planner')
       const response = await axios.get(`https://mvp-api-5dvjwdatfq-ew.a.run.app/planner/${userToken}/${subject}`)
       console.log('got planner')
       return response
     }
+  }
 
   return null
 }
@@ -43,10 +45,12 @@ export const getUserPlanner = async (user, subject) => {
 export const deleteUserPlanner = async (user, subject) => {
 
   if (user) {
-    const userToken = await user.getIdToken()
-    const response = await axios.delete(`https://mvp-api-5dvjwdatfq-ew.a.run.app/planner/${userToken}/${subject}`)
-    console.log(response)
-    return response
+    if (!user.isAnonymous) {
+      const userToken = await user.getIdToken()
+      const response = await axios.delete(`https://mvp-api-5dvjwdatfq-ew.a.run.app/planner/${userToken}/${subject}`)
+      console.log(response)
+      return response
+    }
   }
 
   return null
@@ -66,14 +70,16 @@ export const getDefaultPlanner = async (subject, config) => {
 export const setUserPlanner = async (user, subject, config) => {
 
   if (user) {
-    const userToken = await user.getIdToken()
-    const response = await axios.post(
-        `https://mvp-api-5dvjwdatfq-ew.a.run.app/planner/${userToken}/${subject}`, {
-            config: JSON.stringify(config),
-    }, { headers: { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*' } }
-    )
+    if (!user.isAnonymous) {
+      const userToken = await user.getIdToken()
+      const response = await axios.post(
+          `https://mvp-api-5dvjwdatfq-ew.a.run.app/planner/${userToken}/${subject}`, {
+              config: JSON.stringify(config),
+      }, { headers: { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*' } }
+      )
 
-    return response
+      return response
+    }
   }
 
   return null
@@ -82,14 +88,16 @@ export const setUserPlanner = async (user, subject, config) => {
 export const updateUserPlanner = async (user, subject, changes) => {
 
   if (user) {
-    const userToken = await user.getIdToken()
+    if (!user.isAnonymous) {
+      const userToken = await user.getIdToken()
 
-    await axios.put(
-        `https://mvp-api-5dvjwdatfq-ew.a.run.app/planner/${userToken}/${subject}`,
-        JSON.stringify({changes: changes }),
-      { headers: { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*' } }
-    )
+      await axios.put(
+          `https://mvp-api-5dvjwdatfq-ew.a.run.app/planner/${userToken}/${subject}`,
+          JSON.stringify({changes: changes }),
+        { headers: { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*' } }
+      )
 
+    }
   }
 
   return null
